requirements.txt
google-adk>=1.0.0
google-cloud-aiplatform>=1.91.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
cryptography>=44.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
